Provided convenience HID Report descriptor method for joystick/gamepad
- Added createJoystickReportDescriptor to hid_report_descriptor.py
- Renamed usb_hid_report_descriptor to hid_report_descriptor
