+++ joystick.py
@@ -2,6 +2,7 @@ from smbus import SMBus
 import RPi.GPIO as GPIO
 import time
 
+
 class Joystick:
     # Based on https://github.com/pimoroni/explorer-hat/blob/master/library/explorerhat/ads1015.py
