--- sdp_record_gamepad.xml
@@ -1,114 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<record>
-	<attribute id="0x0001">
-		<sequence>
-			<uuid value="0x1124" />
-		</sequence>
-	</attribute>
-	<attribute id="0x0004">
-		<sequence>
-			<sequence>
-				<uuid value="0x0100" />
-				<uint16 value="0x0011" />
-			</sequence>
-			<sequence>
-				<uuid value="0x0011" />
-			</sequence>
-		</sequence>
-	</attribute>
-	<attribute id="0x0005">
-		<sequence>
-			<uuid value="0x1002" />
-		</sequence>
-	</attribute>
-	<attribute id="0x0006">
-		<sequence>
-			<uint16 value="0x656e" />
-			<uint16 value="0x006a" />
-			<uint16 value="0x0100" />
-		</sequence>
-	</attribute>
-	<attribute id="0x0009">
-		<sequence>
-			<sequence>
-				<uuid value="0x1124" />
-				<uint16 value="0x0100" />
-			</sequence>
-		</sequence>
-	</attribute>
-	<attribute id="0x000d">
-		<sequence>
-			<sequence>
-				<sequence>
-					<uuid value="0x0100" />
-					<uint16 value="0x0013" />
-				</sequence>
-				<sequence>
-					<uuid value="0x0011" />
-				</sequence>
-			</sequence>
-		</sequence>
-	</attribute>
-	<attribute id="0x0100">
-		<text value="A Virtual Gamepad Controller" />
-	</attribute>
-	<attribute id="0x0101">
-		<text value="Keyboard > BT Gamepad" />
-	</attribute>
-	<attribute id="0x0102">
-		<text value="GCC" />
-	</attribute>
-	<attribute id="0x0200">
-		<uint16 value="0x0100" />
-	</attribute>
-	<attribute id="0x0201">
-		<uint16 value="0x0111" />
-	</attribute>
-	<attribute id="0x0202">
-		<uint8 value="0x08" />
-	</attribute>
-	<attribute id="0x0203">
-		<uint8 value="0x00" />
-	</attribute>
-	<attribute id="0x0204">
-		<boolean value="false" />
-	</attribute>
-	<attribute id="0x0205">
-		<boolean value="false" />
-	</attribute>
-	<attribute id="0x0206">
-		<sequence>
-			<sequence>
-				<uint8 value="0x22" />
-				<text encoding="hex" value="05010905a101a103850105091901290215002501950e75018102950175028103a100050109300931093309341581257F750895048102c0c0c0" />
-			</sequence>
-		</sequence>
-	</attribute>
-	<attribute id="0x0207">
-		<sequence>
-			<sequence>
-				<uint16 value="0x0409" />
-				<uint16 value="0x0100" />
-			</sequence>
-		</sequence>
-	</attribute>
-	<attribute id="0x020b">
-		<uint16 value="0x0100" />
-	</attribute>
-	<attribute id="0x020c">
-		<uint16 value="0x0c80" />
-	</attribute>
-	<attribute id="0x020d">
-		<boolean value="true" />
-	</attribute>
-	<attribute id="0x020e">
-		<boolean value="false" />
-	</attribute>
-	<attribute id="0x020f">
-		<uint16 value="0x0640" />
-	</attribute>
-	<attribute id="0x0210">
-		<uint16 value="0x0320" />
-	</attribute>
-</record>
